+++ jovo-clients/jovo-client-web/src/index.ts
@@ -15,9 +15,10 @@ declare global {
 export type {
   CorePlatformRequestJSON as WebRequest,
   CorePlatformResponseJSON as WebResponse,
+  Version,
 } from 'jovo-platform-web';
 // tslint:disable-next-line
-export * from 'jovo-platform-core/dist/src/Interfaces'; // ugly fix but best way to only re-export what is needed
+export * from 'jovo-platform-web/dist/src/Interfaces'; // ugly fix but best way to only re-export what is needed
 
 export * from './types';
 export * from './util';