+++ jovo-clients/jovo-client-web/src/util/index.ts
@@ -1,6 +1,7 @@
 export * from './AudioHelper';
 export * from './Base64Converter';
-export * from './OSHelper';
+export * from './BrowserDetector';
+export * from './OSDetector';
 
 export function delay(amountInMs: number): Promise<void> {
   return new Promise((resolve) => {