+++ jovo-platforms/jovo-platform-alexa/src/index.ts
@@ -87,10 +87,11 @@ declare module 'jovo-core/dist/src/core/BaseApp' {
   interface BaseApp {
     /**
      * Sets alexa handlers
+     * @deprecated use this.setPlatformHandler('Alexa', ...handler) instead
      * @public
      * @param {*} handler
      */
-    setAlexaHandler(...handler: Handler[]): this; // tslint:disable-line
+    setAlexaHandler(...handler: Handler[]): this;
   }
 }