+++ jovo-platforms/jovo-platform-googleassistantconv/src/core/ConversationalActionRequest.ts
@@ -8,6 +8,7 @@ import {
   Capability,
   Context,
   Device,
+  GoogleAssistantDeviceName,
   Handler,
   Home,
   Intent,
@@ -42,8 +43,16 @@ export class ConversationalActionRequest implements JovoRequest {
     return this.session?.id;
   }
 
-  getDeviceName() {
-    return '';
+  getDeviceName(): GoogleAssistantDeviceName {
+    if (this.hasScreenInterface()) {
+      if (this.hasWebBrowserInterface()) {
+        return GoogleAssistantDeviceName.GOOGLE_ASSISTANT_PHONE;
+      } else {
+        return GoogleAssistantDeviceName.GOOGLE_ASSISTANT_SMARTDISPLAY;
+      }
+    } else {
+      return GoogleAssistantDeviceName.GOOGLE_ASSISTANT_SPEAKER;
+    }
   }
 
   getIntentName(): string | undefined {