+++ jovo-platforms/jovo-platform-googleassistantconv/src/core/GoogleAssistantRequestBuilder.ts
@@ -1,5 +1,6 @@
 import { Inputs, RequestBuilder } from 'jovo-core';
 import { ConversationalActionRequest } from './ConversationalActionRequest';
+import * as path from 'path';
 const samples: { [key: string]: string } = {
   LaunchRequest: './../../sample-request-json/LAUNCH.json',
   IntentRequest: './../../sample-request-json/IntentRequest.json',
@@ -9,12 +10,12 @@ const samples: { [key: string]: string } = {
 export class GoogleAssistantRequestBuilder implements RequestBuilder<ConversationalActionRequest> {
   type = 'GoogleAction';
 
-  launch(json?: object): Promise<ConversationalActionRequest> {
-    //tslint:disable-line
-    return this.launchRequest(json);
+  async launch(json?: object): Promise<ConversationalActionRequest> {
+    // tslint:disable-line
+    return await this.launchRequest(json);
   }
-  async intent(json?: object): Promise<ConversationalActionRequest>; //tslint:disable-line
-  async intent(name?: string, inputs?: any): Promise<ConversationalActionRequest>; //tslint:disable-line
+  async intent(json?: object): Promise<ConversationalActionRequest>; // tslint:disable-line
+  async intent(name?: string, inputs?: any): Promise<ConversationalActionRequest>; // tslint:disable-line
   // tslint:disable-next-line
   async intent(obj?: any, inputs?: any): Promise<ConversationalActionRequest> {
     if (typeof obj === 'string') {
@@ -36,49 +37,80 @@ export class GoogleAssistantRequestBuilder implements RequestBuilder<Conversatio
   }
 
   async launchRequest(json?: object): Promise<ConversationalActionRequest> {
+    //tslint:disable-line
     if (json) {
       return ConversationalActionRequest.fromJSON(json);
     } else {
-      const request = JSON.stringify(require(samples['LaunchRequest']));
-      return ConversationalActionRequest.fromJSON(JSON.parse(request));
+      // const request = await fsreadFile(getJsonFilePath('LAUNCH'], 'utf8');
+      const request = JSON.stringify(require(getJsonFilePath('LaunchRequest')));
+      return ConversationalActionRequest.fromJSON(JSON.parse(request)).setTimestamp(
+        new Date().toISOString(),
+      );
     }
   }
   async intentRequest(json?: object): Promise<ConversationalActionRequest> {
+    // tslint:disable-line
     if (json) {
       return ConversationalActionRequest.fromJSON(json);
     } else {
-      const request = JSON.stringify(require(samples['IntentRequest']));
-      return ConversationalActionRequest.fromJSON(JSON.parse(request));
+      // const request = await fsreadFile(getJsonFilePath('LAUNCH'], 'utf8');
+      const request = JSON.stringify(require(getJsonFilePath('IntentRequest')));
+      return ConversationalActionRequest.fromJSON(JSON.parse(request)).setTimestamp(
+        new Date().toISOString(),
+      );
     }
   }
 
-  async rawRequest(json: object): Promise<ConversationalActionRequest> {
-    //tslint:disable-line
+  async rawRequest(json: object) {
+    // tslint:disable-line
     return ConversationalActionRequest.fromJSON(json);
   }
 
-  async rawRequestByKey(key: string): Promise<ConversationalActionRequest> {
-    const request = JSON.stringify(require(samples[key]));
+  async rawRequestByKey(key: string) {
+    const request = JSON.stringify(require(getJsonFilePath(key)));
     return ConversationalActionRequest.fromJSON(JSON.parse(request));
   }
 
-  // TODO:
-  async audioPlayerRequest(json?: object): Promise<ConversationalActionRequest> {
+  async audioPlayerRequest(json?: object) {
+    // tslint:disable-line
     if (json) {
       return ConversationalActionRequest.fromJSON(json);
     } else {
-      // const request = await fsreadFile(samples['LAUNCH'], 'utf8');
-      const request = JSON.stringify(require(samples['IntentRequest1']));
-      return ConversationalActionRequest.fromJSON(JSON.parse(request));
+      const request = JSON.stringify(require(getJsonFilePath('AudioPlayer.PlaybackStarted')));
+      return ConversationalActionRequest.fromJSON(JSON.parse(request)).setTimestamp(
+        new Date().toISOString(),
+      );
     }
+  }
+  /**
+   * End
+   * @param {object|string} json
+   * @return {SessionEndedRequest}
+   */
+  // tslint:disable-next-line
+  async end(json?: any) {
+    if (json) {
+      return ConversationalActionRequest.fromJSON(json);
+    } else {
+      const request = JSON.stringify(require(getJsonFilePath('EndRequest')));
+      return ConversationalActionRequest.fromJSON(JSON.parse(request)).setTimestamp(
+        new Date().toISOString(),
+      );
+    }
+  }
+}
+function getJsonFilePath(key: string, version = 'v1'): string {
+  let folder = './../../../';
+
+  if (process.env.NODE_ENV === 'UNIT_TEST') {
+    folder = './../../';
   }
 
-  async end(json?: object): Promise<ConversationalActionRequest> {
-    if (json) {
-      return ConversationalActionRequest.fromJSON(json);
-    } else {
-      const request = JSON.stringify(require(samples['EndRequest']));
-      return ConversationalActionRequest.fromJSON(JSON.parse(request));
-    }
+  const fileName = samples[key];
+
+  if (!fileName) {
+    throw new Error(`${key} Can't find file.`);
   }
+
+  return path.join(folder, 'sample-request-json', version, fileName);
 }