+++ jovo-platforms/jovo-platform-googleassistantconv/src/modules/ConversationalActionsCore.ts
@@ -73,8 +73,23 @@ export class ConversationalActionsCore implements Plugin {
     };
 
     GoogleAction.prototype.addTypeOverrides = function (typeOverrides: TypeOverride[]) {
+      let currentTypeOverrides: TypeOverride[] = _get(
+        this.$output,
+        'GoogleAssistant.typeOverrides',
+        [],
+      );
+      currentTypeOverrides = currentTypeOverrides.concat(typeOverrides);
+      _set(this.$output, 'GoogleAssistant.typeOverrides', currentTypeOverrides);
+      return this;
+    };
+
+    GoogleAction.prototype.setTypeOverrides = function (typeOverrides: TypeOverride[]) {
       _set(this.$output, 'GoogleAssistant.typeOverrides', typeOverrides);
       return this;
+    };
+
+    GoogleAction.prototype.showBasicCard = function (basicCard: Card) {
+      return this.addCard(basicCard);
     };
   }
 
@@ -399,11 +414,18 @@ export class ConversationalActionsCore implements Plugin {
       });
     }
 
-    if (output.GoogleAssistant?.suggestions) {
+    const suggestions = output.GoogleAssistant?.suggestions;
+    const quickReplies = output.quickReplies;
+    if (suggestions?.length || quickReplies?.length) {
+      const newSuggestions = suggestions?.length
+        ? suggestions
+        : quickReplies!.map((quickReply) => ({
+            title: typeof quickReply !== 'string' ? quickReply.value : quickReply,
+          }));
       _set(
         googleAction.$response as ConversationalActionResponse,
         'prompt.suggestions',
-        output.GoogleAssistant.suggestions,
+        newSuggestions,
       );
     }
 
@@ -478,6 +500,7 @@ export class ConversationalActionsCore implements Plugin {
         id: request.session?.id!,
         params: {
           _JOVO_SESSION_: googleAction.$conversationalSession,
+          _JOVO_STATE_: null,
           ...googleAction.$session.$data,
         },
       };
@@ -496,6 +519,14 @@ export class ConversationalActionsCore implements Plugin {
         googleAction.$response as ConversationalActionResponse,
         'scene.next.name',
         output.GoogleAssistant.nextScene,
+      );
+    }
+
+    if (output.GoogleAssistant?.expected) {
+      _set(
+        googleAction.$response as ConversationalActionResponse,
+        'expected',
+        output.GoogleAssistant.expected,
       );
     }
   }