+++ jovo-platforms/jovo-platform-web/src/Interfaces.ts
@@ -1,0 +1,2 @@
+// tslint:disable-next-line
+export * from 'jovo-platform-core/dist/src/Interfaces';