+++ jovo-platforms/jovo-platform-web/src/index.ts
@@ -1,6 +1,7 @@
 import { Action, Config } from 'jovo-platform-core';
 import { WebApp } from './core/WebApp';
 
+export * from './Interfaces';
 export * from 'jovo-platform-core';
 export * from './WebPlatform';
 
@@ -27,6 +28,7 @@ declare module 'jovo-core/dist/src/Interfaces' {
   }
 
   interface AppPlatformConfig extends AppWebPlatformConfig {}
+
   interface ExtensiblePluginConfigs extends AppWebPlatformConfig {}
 }