+++ jovo-platforms/jovo-platform-googleassistantconv/CHANGELOG.md
@@ -2,7 +2,21 @@
 
 Current version might be higher than the latest changes displayed below because of updates of dependencies.
 
+##### 2020-12-03 [3.3.0]
+- [#871](https://github.com/jovotech/jovo-framework/pull/871) Add missing and broken Google Conversational Action features ([@aswetlow](https://github.com/aswetlow))
 
+##### 2020-11-20 [3.2.4]
+- [#862](https://github.com/jovotech/jovo-framework/pull/862) Fix removeState ([@aswetlow](https://github.com/aswetlow))
+
+##### 2020-11-16 [3.2.3]
+- Adds Interactive Canvas to Google Conversational Actions ([@aswetlow](https://github.com/aswetlow))
+
+##### 2020-11-10 [3.2.2]
+- [#856](https://github.com/jovotech/jovo-framework/pull/856) Fixes several Google Conversational Actions issues ([@aswetlow](https://github.com/aswetlow))
+
+
+##### 2020-11-05 [3.2.1]
+Fix missing locale in push notifications object
 
 ##### 2020-09-29 [3.1.3]
 - [#831](https://github.com/jovotech/jovo-framework/pull/831) :recycle: Enhance Google AssisConversational Actions ([@maswetlow](https://github.com/aswetlow))